chore: creation of 'service-mesh' module
