--- src/services/index.ts
@@ -1,2 +1,0 @@
-export { constant, get, container, createInjections, build, merge } from './utils'
-export { TOKENS, services } from './production'
