+++ src/utilities/index.ts
@@ -1,5 +1,6 @@
-import { TOKENS } from '@/services/tokens'
+import { TOKENS } from '@/app/kuma'
 import { createInjections } from '@/services/utils'
+
 export { useEnv, useI18n } from '@/app/application'
 export { useRouter } from '@/app/vue'
